// src/app/api/model/consent-status-dto-out.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ConsentRecordSummary } from './consent-record-summary';


export interface ConsentStatusDtoOut { 
    daysToAcceptNewTerms?: number;
    newestConsentsAccepted?: boolean;
    records?: Array<ConsentRecordSummary>;
}