// src/app/api/model/legal-document-type.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * LegalDocumentType enum name
 */
export enum LegalDocumentType {
    COOKIE_POLICY = 'COOKIE_POLICY',
    TERMS_OF_SERVICE = 'TERMS_OF_SERVICE',
    PRIVACY_POLICY = 'PRIVACY_POLICY',
    SUBSCRIPTION_ACTIVATION_CONSENT = 'SUBSCRIPTION_ACTIVATION_CONSENT',
    DATA_RETENTION_POLICY = 'DATA_RETENTION_POLICY'
}